requirements.txt
pyYAML
requests
Adafruit_DHT